python-uv/pyproject.toml
[project]
name = "python-uv"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dateutil>=2.9.0.post0",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
    "types-python-dateutil>=2.9.0.20241206",
]
